backend/requirements.txt
wheel==0.37.1
alembic==1.8.1
cryptography==37.0.2
FastAPI>=0.85.1,<0.86.0
mysqlclient==2.1.1
passlib[bcrypt]==1.7.4
pydantic>=1.9.0,<2.0.0
pydantic[email]
python-dotenv==0.19.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
sqlmodel==0.0.8
typer==0.4.2
Uvicorn==0.18.2
